retoactive updates.

+++ includes/Admin/AllFormsTable.php
@@ -264,12 +264,20 @@ class NF_Admin_AllFormsTable extends WP_List_Table
             || ( isset( $_POST['action2'] ) && $_POST['action2'] == 'bulk-delete' )
         ) {
 
-            $delete_ids = esc_sql( $_POST['bulk-delete'] );
+            // In our file that handles the request, verify the nonce.
+            $nonce = esc_attr( $_REQUEST['_wpnonce'] );
+            if ( ! wp_verify_nonce( $nonce, 'bulk-forms' ) ) {
+                die( 'Go get a life, script kiddies' );
+            }
 
-            // loop over the array of record IDs and delete them
-            foreach ( $delete_ids as $id ) {
+            if( isset( $_POST[ 'bulk-delete' ] ) ) {
+                $delete_ids = esc_sql($_POST['bulk-delete']);
 
-                self::delete_item( absint( $id ) );
+                // loop over the array of record IDs and delete them
+                foreach ($delete_ids as $id) {
+
+                    self::delete_item(absint($id));
+                }
             }
 
             wp_redirect( admin_url( 'admin.php?page=ninja-forms' ) );
